requirements.txt
asn1crypto==1.0.1
cffi==1.12.3
cryptography==2.7
Django==2.2.6
djangorestframework==3.10.3
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
